separate messaging and scriptSource provider
