+++ src/notebooks/controllers/ipywidgets/message/ipyWidgetMessageDispatcherFactory.ts
@@ -5,11 +5,11 @@
 
 import { inject, injectable } from 'inversify';
 import { Event, EventEmitter, NotebookDocument } from 'vscode';
-import { IDisposable, IDisposableRegistry } from '../../../platform/common/types';
-import { IPyWidgetMessages } from '../../../messageTypes';
-import { IKernel, IKernelProvider } from '../../../kernels/types';
+import { IDisposable, IDisposableRegistry } from '../../../../platform/common/types';
+import { IPyWidgetMessages } from '../../../../messageTypes';
+import { IKernel, IKernelProvider } from '../../../../kernels/types';
 import { IPyWidgetMessageDispatcher } from './ipyWidgetMessageDispatcher';
-import { IIPyWidgetMessageDispatcher, IPyWidgetMessage } from './types';
+import { IIPyWidgetMessageDispatcher, IPyWidgetMessage } from '../types';
 
 /**
  * This just wraps the iPyWidgetMessageDispatcher class.
