+++ src/notebooks/controllers/ipywidgets/scriptSourceProvider/ipyWidgetScriptManagerFactory.web.ts
@@ -3,11 +3,11 @@
 // Licensed under the MIT License.
 
 import { injectable, inject } from 'inversify';
-import { IFileSystem } from '../../../platform/common/platform/types';
-import { IExtensionContext, IHttpClient } from '../../../platform/common/types';
-import { IKernel } from '../../../kernels/types';
+import { IFileSystem } from '../../../../platform/common/platform/types';
+import { IExtensionContext, IHttpClient } from '../../../../platform/common/types';
+import { IKernel } from '../../../../kernels/types';
 import { RemoteIPyWidgetScriptManager } from './remoteIPyWidgetScriptManager';
-import { IIPyWidgetScriptManager, IIPyWidgetScriptManagerFactory } from './types';
+import { IIPyWidgetScriptManager, IIPyWidgetScriptManagerFactory } from '../types';
 
 /**
  * Determines the IPyWidgetScriptManager for use in a web environment
