+++ src/notebooks/controllers/ipywidgets/scriptSourceProvider/nbExtensionsPathProvider.web.ts
