+++ src/test/datascience/ipywidgets/localWidgetScriptSourceProvider.unit.test.ts
@@ -4,7 +4,7 @@ import { assert } from 'chai';
 import { anything, instance, mock, when } from 'ts-mockito';
 import { Uri } from 'vscode';
 import { IKernel } from '../../../platform/../kernels/types';
-import { LocalWidgetScriptSourceProvider } from '../../../notebooks/controllers/ipywidgets/localWidgetScriptSourceProvider.node';
+import { LocalWidgetScriptSourceProvider } from '../../../notebooks/controllers/ipywidgets/scriptSourceProvider/localWidgetScriptSourceProvider.node';
 import {
     ILocalResourceUriConverter,
     IWidgetScriptSourceProvider,
